+++ packages/bus-core/src/service-bus/index.ts
@@ -1,3 +1,4 @@
 export * from './bus'
 export * from './bus-instance'
 export * from './bus-state'
+export * from './bus-configuration'
