+++ packages/bus-core/src/util/index.ts
@@ -3,3 +3,4 @@ export * from './class-constructor'
 export * from './assert-unreachable'
 export * from './typed-emitter'
 export * from './core-dependencies'
+export * from './middleware'
